🐛 修复使用纯数字时验证错误的问题
